move bridge to internal/bridge
