+++ drivers/mac/notification.go
@@ -6,7 +6,7 @@ import (
 	"github.com/murlokswarm/app"
 
 	"github.com/google/uuid"
-	"github.com/murlokswarm/app/bridge"
+	"github.com/murlokswarm/app/internal/bridge"
 	"github.com/murlokswarm/app/internal/core"
 )
