+++ internal/bridge/go.go
