+++ internal/bridge/go_test.go
