+++ internal/bridge/platform.go
