+++ internal/bridge/platform_test.go
